+++ ocdsmerge/__init__.py
@@ -1,1 +1,1 @@
-from .merge import APPEND, MERGE_BY_POSITION, RAISE, WARN, get_merge_rules, merge, merge_versioned  # noqa
+from .merge import APPEND, MERGE_BY_POSITION, get_merge_rules, merge, merge_versioned  # noqa